pyproject.toml
[project]
name = "product-recommender"
version = "0.1.0"
description = ""
authors = [
    {name = "Satish Nande",email = "satishnande23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "azure-storage-blob (>=12.25.1,<13.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]

[tool.poetry]
packages = [{include = "product_recommender", from = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.9"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
pandas-stubs = "^2.2.3.250308"
ipykernel = "^6.29.5"
nbformat = "^5.10.4"
ipywidgets = "^8.1.7"
types-requests = "^2.32.4.20250611"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "dist",
    "build",
    "docs",
]

[tool.ruff.lint]
extend-select = ["E", "F", "W", "I", "B", "UP", "SIM", "N", "D"]
ignore = []


[tool.ruff.lint.isort]
known-first-party = ["product_recommender"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-annotations]
mypy-init-return = false
suppress-none-returning = true
allow-star-arg-any = false
ignore-fully-untyped = false


[tool.mypy]
python_version = "3.12"
files = ["src/product_recommender", "tests", "pages"]
strict = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true